+++ src/app/core/register2/register2.backend-validators.ts
@@ -25,7 +25,8 @@ interface HasFormAdapters {
     StepB: UntypedFormGroup,
     StepC: UntypedFormGroup,
     StepC2: UntypedFormGroup,
-    StepD: UntypedFormGroup
+    StepD: UntypedFormGroup,
+    isReactivation?: boolean
   ): RegisterForm
 }
 
@@ -164,7 +165,7 @@ export function Register2BackendValidatorMixin<
       StepC: UntypedFormGroup,
       StepC2: UntypedFormGroup,
       StepD: UntypedFormGroup,
-
+      isReactivation?: boolean,
       type?: 'shibboleth'
     ): Observable<RegisterForm> {
       const registerForm = this.formGroupToFullRegistrationForm(
@@ -174,6 +175,7 @@ export function Register2BackendValidatorMixin<
         StepC2,
         StepD
       )
+      registerForm.isReactivation = isReactivation
       return this._http
         .post<RegisterForm>(
           `${runtimeEnvironment.API_WEB}register.json`,